# headless/headless-commerce/headless-commerce-admin-inventory-impl/rest-openapi.yaml
components:
    schemas:
        Error:
            properties:
                errorCode:
                    description: Internal error code mapping
                    example: 996
                    readOnly: true
                    type: integer
                errorDescription:
                    example: Unable to find currency. Currency code should be expressed with
                        3-letter ISO 4217 format.
                    readOnly: true
                    type: string
                message:
                    example: No CommerceCurrency exists with the key {groupId=41811, code=US
                        Dollar}
                    readOnly: true
                    type: string
                status:
                    description: HTTP Status code
                    example: 404
                    readOnly: true
                    type: integer
            required:
                - errorCode
                - errorDescription
                - message
                - status
            type: object
        Warehouse:
            properties:
                active:
                    example: true
                    type: boolean
                city:
                    example: Diamond Bar
                    type: string
                countryISOCode:
                    example: US
                    type: string
                description:
                    example: right stairs, first room on the left
                    type: string
                externalReferenceCode:
                    example: AB-34098-789-N
                    type: string
                id:
                    example: 30130
                    format: int64
                    minimum: 0
                    type: integer
                items:
                    items:
                        $ref: '#/components/schemas/WarehouseItem'
                    type: array
                latitude:
                    example: 33.9976884
                    format: double
                    type: number
                longitude:
                    example: -117.8144595
                    format: double
                    type: number
                name:
                    example: Alessio Antonio Rendina
                    type: string
                regionISOCode:
                    example: CA
                    type: string
                street1:
                    example: 1400 Montefino Ave
                    type: string
                street2:
                    example: 1st floor
                    type: string
                street3:
                    example: suite 200
                    type: string
                type:
                    type: string
                zip:
                    example: "91765"
                    type: string
            type: object
        WarehouseItem:
            properties:
                id:
                    example: 30130
                    format: int64
                    minimum: 0
                    type: integer
                modifiedDate:
                    example: 2017-07-21
                    format: date
                    type: string
                quantity:
                    example: 1
                    format: int32
                    minimum: 0
                    type: integer
                reservedQuantity:
                    example: 2
                    format: int32
                    minimum: 0
                    type: integer
                sku:
                    example: SKU
                    type: string
                warehouseExternalReferenceCode:
                    example: AB-34098-789-N
                    type: string
                warehouseId:
                    example: 30030
                    format: int64
                    minimum: 0
                    type: integer
            type: object
    securitySchemes:
        BasicAuth:
            scheme: basic
            type: http
        BearerAuth:
            scheme: bearer
            type: http
        OAuth2:
            flows:
                authorizationCode:
                    authorizationUrl: /oauth/authorize
                    scopes:
                        CommerceOpenApiAdmin.admin: Grants access to admin operations
                        CommerceOpenApiAdmin.read: Grants read access
                        CommerceOpenApiAdmin.write: Grants write access
                    tokenUrl: /oauth/token
            type: oauth2
info:
    contact:
        email: team-commerce@liferay.com
        name: Commerce Team
    description: Liferay Commerce Admin Inventory API
    title: Liferay Commerce Admin Inventory API
    version: v1.0
openapi: 3.0.0
paths:
...
    /warehouses/:
        get:
            operationId: getWarehousesPage
            parameters:
                - in: query
                  name: page
                  required: false
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  required: false
                  schema:
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/Warehouse'
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: '#/components/schemas/Warehouse'
                                type: array
                    description: Successful operation
                400:
                    description: Invalid input
                401:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - CommerceOpenApiAdmin.read
            summary: Gets a List of Warehouses.
            tags:
                - Warehouse
        post:
            operationId: postWarehous
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Warehouse'
                    application/xml:
                        schema:
                            $ref: '#/components/schemas/Warehouse'
                required: true
            responses:
                201:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Warehouse'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Warehouse'
                    description: Created
                202:
                    description: Accepted - Async
                204:
                    description: Updated
                400:
                    description: Invalid input
                401:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - CommerceOpenApiAdmin.write
            summary: Creates or updates a Warehouse.
            tags:
                - Warehouse
...
    /warehouses/by-externalReferenceCode/{externalReferenceCode}:
        delete:
            operationId: deleteWarehousByExternalReferenceCode
            parameters:
                - in: path
                  name: externalReferenceCode
                  required: true
                  schema:
                      description: The resource external reference code
                      example: AB-34098-789-N
                      minLength: 1
                      type: string
            responses:
                204:
                    description: No content
                401:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Authentication information is missing or invalid
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - CommerceOpenApiAdmin.write
            summary: Deletes a Warehouse by external reference code.
            tags:
                - Warehouse
        get:
            operationId: getWarehousByExternalReferenceCode
            parameters:
                - in: path
                  name: externalReferenceCode
                  required: true
                  schema:
                      description: The resource external reference code
                      example: AB-34098-789-N
                      minLength: 1
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Warehouse'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Warehouse'
                    description: Successful operation
                400:
                    description: Invalid input
                401:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - CommerceOpenApiAdmin.read
            summary: Gets a Warehouse by external reference code.
            tags:
                - Warehouse
        patch:
            operationId: patchWarehousByExternalReferenceCode
            parameters:
                - in: path
                  name: externalReferenceCode
                  required: true
                  schema:
                      description: The resource external reference code
                      example: AB-34098-789-N
                      minLength: 1
                      type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Warehouse'
                    application/xml:
                        schema:
                            $ref: '#/components/schemas/Warehouse'
                required: true
            responses:
                202:
                    description: Async
                204:
                    description: Updated
                400:
                    description: Invalid input
                401:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - CommerceOpenApiAdmin.write
            summary: Updates a Warehouse by external reference code.
            tags:
                - Warehouse
...
    /warehouses/{id}:
        delete:
            operationId: deleteWarehousId
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      description: The resource Id
                      example: 30130
                      format: int64
                      minimum: 0
                      type: integer
            responses:
                204:
                    description: No content
                401:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Authentication information is missing or invalid
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - CommerceOpenApiAdmin.write
            summary: Deletes a Warehouse by ID.
            tags:
                - Warehouse
        get:
            operationId: getWarehousId
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      description: The resource Id
                      example: 30130
                      format: int64
                      minimum: 0
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Warehouse'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Warehouse'
                    description: Successful operation
                400:
                    description: Invalid input
                401:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - CommerceOpenApiAdmin.read
            summary: Gets a Warehouse by ID.
            tags:
                - Warehouse
        patch:
            operationId: patchWarehousId
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      description: The resource Id
                      example: 30130
                      format: int64
                      minimum: 0
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Warehouse'
                    application/xml:
                        schema:
                            $ref: '#/components/schemas/Warehouse'
                required: true
            responses:
                202:
                    description: Async
                204:
                    description: Updated
                400:
                    description: Invalid input
                401:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - CommerceOpenApiAdmin.write
            summary: Updates a Warehouse by ID.
            tags:
                - Warehouse
servers:
    - description: Liferay Commerce API server
      url: https://commerce.{environment}.liferay.com/commerce-admin-inventory/{basePath}
      variables:
          basePath:
              default: v1.0
          environment:
              default: api
              enum:
                  - api
                  - api.dev
                  - api.uat
    - description: Liferay Commerce Local Development (Hosts file)
      url: '{protocol}://{host}:{port}/o/commerce-admin-inventory/{basePath}'
      variables:
          basePath:
              default: v1.0
          host:
              default: test.dev
          port:
              default: "8080"
              enum:
                  - "443"
                  - "8080"
                  - "8443"
          protocol:
              default: http
              enum:
                  - http
                  - https
    - description: Liferay Commerce Local Development with Ngrok
      url: '{protocol}://{host}.ngrok.io/o/commerce-admin-inventory/{basePath}'
      variables:
          basePath:
              default: v1.0
          host:
              default: commerce
          protocol:
              default: http
              enum:
                  - http
                  - https
...